Remove files that are not being used anymore.
