--- docs/explainReduxStructure.md
@@ -1,1 +1,0 @@
-\
