+++ src/common/context.js
@@ -1,4 +1,3 @@
 import React from "react";
-
 export const DeviceProvider = React.createContext({});
 export const LanguageProvider = React.createContext({});
