--- src/components/Course/Content/ExerciseList/Exercise/CollapseArrow/styles.scss
@@ -1,11 +1,0 @@
-.ng-collapse-arrow {
-  position: absolute;
-  top: 10px;
-  right: 12px;
-
-  > span {
-    font-size: 24px;
-    width: 24px;
-    height: 12px;
-  }
-}
