--- src/components/Course/Content/ExerciseList/Exercise/styles.scss
@@ -1,59 +1,0 @@
-.ng-exercise {
-  position: relative;
-  width: 300px;
-  padding: 8px;
-  background: #ffffff;
-  text-decoration: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
-    0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12);
-  align-self: center;
-  cursor: pointer;
-
-  .content {
-    display: flex;
-    align-items: center;
-    padding-right: 36px;
-  }
-
-  .child-exercise {
-    padding-top: 16px;
-    padding-left: 32px;
-  }
-
-  .logo {
-    width: 34px;
-    height: 34px;
-    padding-top: 24px;
-    margin-left: 10px;
-    margin-right: 10px;
-    margin-top: -20px;
-
-    > .selected {
-      fill: #f05f40;
-    }
-  }
-
-  .child-exercise-logo {
-    width: 24px;
-    height: 24px;
-  }
-
-  > .title {
-    width: 70%;
-    margin-left: 8px;
-  }
-}
-
-.ng-exercise-selected {
-  width: 316px;
-  color: #d55f31;
-}
-
-.ng-exercise-child {
-  margin-top: 16px;
-  margin-bottom: 16px;
-  padding-bottom: 16px;
-}
