--- src/components/Course/Content/ExerciseList/styles.scss
@@ -1,29 +1,0 @@
-@import "../../../../sass/colors.scss";
-@import "../../../../sass/mixins";
-
-.ng-exercise-list {
-  display: flex;
-  flex-direction: column;
-  @include breakpoint-mobile {
-    display: none;
-  }
-
-  > .enroll {
-    width: 348px;
-    height: 24px;
-    margin-bottom: 32px;
-    padding: 8px;
-    background-color: #f05f40;
-    text-decoration: none;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    align-self: center;
-    color: white;
-    font-size: 16px;
-    background-color: yellowgreen;
-    box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.2),
-      0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12);
-    cursor: pointer;
-  }
-}
