--- src/components/Course/Content/HiddenContent/styles.scss
@@ -1,24 +1,0 @@
-#hiddenContent {
-  &__container {
-  }
-
-  &__control {
-    cursor: pointer;
-    font-weight: 600;
-    margin-top: 15px;
-    border: 2px solid #111;
-    padding: 2px;
-    width: 130px;
-    user-select: none;
-  }
-
-  &__main {
-    &--hide {
-      visibility: hidden;
-    }
-
-    &--show {
-      visibility: visible;
-    }
-  }
-}
