--- src/components/Course/constant.js
@@ -1,10 +1,0 @@
-export const COURSE_CARD_COLORS = [
-  "#FFA4A4",
-  "#A4C5FF",
-  "#CBE0C4",
-  "#B7A4FF",
-  "#A4C5EE",
-  "#CBE0B4",
-  "#FFA4C4",
-  "#B7A4GG",
-];
