--- src/components/Course/styles.scss
@@ -1,13 +1,0 @@
-@import "../../sass/colors.scss";
-@import "../../sass/mixins.scss";
-
-.ng-course {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #f6f8fa;
-
-  h2 {
-    margin-bottom: 32px;
-  }
-}
