+++ src/components/PathwayExercise/ExerciseContent/index.js
@@ -35,10 +35,8 @@ import CourseEnroll from "../../BatchClassComponents/EnrollInCourse/EnrollInCour
 import DoubtClassExerciseComponent from "../../BatchClassComponents/DoubtClassExerciseComponent";
 import RevisionClassEnroll from "../../BatchClassComponents/Revision/RevisionClassEnroll";
 import { actions as upcomingBatchesActions } from "../..//PathwayCourse/redux/action";
-import { actions as upcomingClassActions } from "../../PathwayCourse/redux/action";
+// import { actions as upcomingClassActions } from "../../PathwayCourse/redux/action";
 import ClassTopic from "../ClassTopic/ClassTopic";
-// import { Container, Box, Typography, Button, Grid } from "@mui/material";
-import languageMap from "../../../pages/CourseContent/languageMap";
 import ExerciseContentLoading from "./ExerciseContentLoading";
 const createVisulizeURL = (code, lang, mode) => {
   // only support two languages for now
