+++ src/components/PathwayExercise/index.js
@@ -14,10 +14,7 @@ import { useParams } from "react-router-dom";
 import { PATHS, interpolatePath, versionCode } from "../../constant";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick.css";
-
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-// const {languageMap} = require("../../pages/CourseContent/languageMap");
-
 import {
   Container,
   Box,
@@ -28,8 +25,15 @@ import {
   Select,
   MenuItem,
 } from "@mui/material";
-import languageMap from "../../pages/CourseContent/languageMap";
 import CompletionComponent from "./CourseCompletion/CompletionComponent";
+
+const languageMap = {
+  hi: "Hindi",
+  en: "English",
+  te: "Telugu",
+  ta: "Tamil",
+  mr: "Marathi",
+};
 
 const Exercise = ({
   course,
