--- src/pages/Course/styles.scss
@@ -1,31 +1,0 @@
-@import "../../sass/colors.scss";
-@import "../../sass/mixins.scss";
-
-.ng-course-container {
-  box-sizing: border-box;
-  padding-top: 40px;
-  max-width: 80%;
-  margin: auto;
-  @include breakpoint-mobile {
-    max-width: 100%;
-    padding: 0 20px;
-  }
-
-  > h2 {
-    margin: 0px;
-    margin-bottom: 8px;
-    margin-left: 32px;
-    font-family: Rubik;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 30px;
-    line-height: 36px;
-    @include breakpoint-mobile {
-      margin-left: 0px;
-    }
-  }
-
-  @include breakpoint-mobile {
-    margin-top: 24px;
-  }
-}
