--- src/pages/CourseContent/languageMap.js
@@ -1,8 +1,0 @@
-const languageMap = {
-  hi: "Hindi",
-  en: "English",
-  te: "Telugu",
-  ta: "Tamil",
-  mr: "Marathi",
-};
-export default languageMap;
