--- src/pages/CourseContent/styles.scss
@@ -1,78 +1,0 @@
-@import "../../sass/colors.scss";
-@import "../../sass/mixins.scss";
-
-.ng-course-content {
-  width: 80%;
-  margin: auto;
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 200px;
-
-  @include breakpoint-mobile {
-    width: 100%;
-  }
-
-  h2 {
-    text-align: center;
-  }
-
-  > .content {
-    margin-right: 24px;
-    padding: 27px;
-    min-height: 100vh;
-    width: 100%;
-    height: fit-content;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background: #ffffff;
-    box-shadow: 0px 12px 18px rgba(41, 94, 255, 0.05);
-    border-radius: 18px;
-    @include breakpoint-mobile {
-      margin-right: 0px;
-      margin-top: 24px;
-    }
-  }
-
-  > .content > .markdown {
-    width: 80%;
-    padding: 24px;
-    padding-top: 24px;
-    font-family: Ubuntu;
-    box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
-      0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12);
-    margin-bottom: 32px;
-    @include breakpoint-mobile {
-      width: 100%;
-    }
-  }
-
-  > .content > .github-link {
-    color: gray;
-    text-decoration: none;
-    width: 80%;
-    text-align: right;
-  }
-
-  > .content > .arrow-row {
-    padding: 32px 32px;
-    width: 80%;
-    display: flex;
-    justify-content: space-between;
-  }
-  > .content > .lang {
-    left: 30%;
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-  }
-
-  .language-select {
-    padding: 6px 8px;
-    font-size: 16px;
-    width: 130px;
-    background: #ffffff;
-    border-radius: 8px;
-    border: 2px solid #d55f31;
-  }
-}
