--- src/pages/Partner/styles.js
@@ -1,74 +1,0 @@
-import { makeStyles } from "@material-ui/styles";
-
-const useStyles = makeStyles((theme) => ({
-  partnerTopspacing: {
-    paddingTop: "2%",
-  },
-  partnerBottomspacing: {
-    marginBottom: "6%",
-  },
-
-  partnerHrline: {
-    width: "90px",
-    textAlign: "center",
-  },
-  partnerBtn: {
-    width: "50%",
-  },
-  partnerBtn1: {
-    height: "60px",
-    width: "100%",
-    fontSize: "lg",
-  },
-  partnerContainer: {
-    paddingTop: "6%",
-    paddingBottom: "6%",
-  },
-  partnerContainer1: {
-    paddingTop: "20%",
-    paddingBottom: "20%",
-  },
-  partnerImg: {
-    height: "100%",
-    width: "100%",
-    borderRadius: "6px",
-  },
-  partnerTopspacing1: {
-    paddingTop: "4%",
-  },
-  partnerIcon: {
-    marginBottom: "-8px",
-  },
-  partnerCard: {
-    height: "95%",
-  },
-  link: {
-    paddingLeft: "20px",
-    paddingTop: "10px",
-    color: "green",
-    textDecorationLine: "none",
-  },
-  link1: {
-    textDecorationLine: "none",
-    color: "white",
-  },
-  link2: {
-    textDecorationLine: "none",
-    color: "black",
-  },
-  partnerCardContainer: {
-    height: "60px",
-  },
-  partnerCardContainer1: {
-    height: "50px",
-  },
-  partnerIconSize: {
-    padding: "4px",
-    backgroundColor: "white",
-    fontSize: "28px",
-    color: " #48A145",
-    borderRadius: "100px",
-    border: "1px solid #48A145",
-  },
-}));
-export default useStyles;
